+++ example/lib/main.dart
@@ -7,7 +7,9 @@ Future<void> main() async {
   WidgetsFlutterBinding.ensureInitialized();
 
   /// initialize theme mode builder
-  await ThemeModeBuilderConfig.ensureInitialized();
+  await ThemeModeBuilderConfig.ensureInitialized(
+    subDir: "Theme Mode Builder Example",
+  );
 
   /// Runs the app :)
   runApp(ThemeModeBuilderExampleApp());